Split out api routing and handlers
